+++ MudBlazor.Extensions/Components/MudExFileDisplay.razor.cs
@@ -8,6 +8,8 @@ using MudBlazor.Extensions.Services;
 using MudBlazor.Extensions.Components.ObjectEdit;
 using MudBlazor.Extensions.Core;
 using MudBlazor.Extensions.Helper.Internal;
+using Nextended.Core.Extensions;
+using Nextended.Core.Helper;
 
 namespace MudBlazor.Extensions.Components;
 
@@ -19,6 +21,7 @@ public partial class MudExFileDisplay : IMudExFileDisplayInfos
 
     #region private fields
 
+    private DynamicComponent _currentFileDisplay;
     private bool _internalCall;
     private bool _isNativeRendered;
     private string _id = Guid.NewGuid().ToString();
@@ -42,6 +45,12 @@ public partial class MudExFileDisplay : IMudExFileDisplayInfos
     /// </summary>
     //[Parameter, SafeCategory("Appearance")]
     public string StatusText { get; private set; }
+
+    /// <summary>
+    /// Specify types of IMudExFileDisplay that should be ignored
+    /// </summary>
+    [Parameter]
+    public Type[] IgnoredRenderControls { get; set; }
 
     /// <summary>
     /// How to handle the stream url
@@ -302,6 +311,9 @@ public partial class MudExFileDisplay : IMudExFileDisplayInfos
         var updateRequired = (parameters.TryGetValue<Stream>(nameof(ContentStream), out var stream) && ContentStream != stream)
                              || (parameters.TryGetValue<string>(nameof(Url), out var url) && Url != url);
 
+        if(updateRequired)
+            _componentForFile = default;
+
         await base.SetParametersAsync(parameters);
 
         if (!updateRequired)
@@ -321,12 +333,19 @@ public partial class MudExFileDisplay : IMudExFileDisplayInfos
     /// <inheritdoc/>
     protected override Task OnParametersSetAsync()
     {
-        _possibleRenderControls = GetServices<IMudExFileDisplay>().Where(c => c.GetType() != GetType() && c.CanHandleFile(this)).ToList();
-        if (ViewDependsOnContentType)
+        _possibleRenderControls = GetServices<IMudExFileDisplay>().Where(c => c.GetType() != GetType() && !Ignored(c) && c.CanHandleFile(this)).ToList();
+        if (ViewDependsOnContentType && _componentForFile == default)
             _componentForFile = GetComponentForFile(_possibleRenderControls.FirstOrDefault(c => c.StartsActive && !ForceNativeRender));
         if (!_internalOverwrite)
             renderInfos = GetRenderInfos();
         return base.OnParametersSetAsync();
+    }
+
+    private bool Ignored(IMudExFileDisplay mudExFileDisplay)
+    {
+        if(IgnoredRenderControls == null || IgnoredRenderControls.Length == 0)
+            return false;
+        return IgnoredRenderControls.Any(t => t.IsInstanceOfType(mudExFileDisplay));
     }
 
 
@@ -336,6 +355,7 @@ public partial class MudExFileDisplay : IMudExFileDisplayInfos
         if (type == null)
             return default;
         var parameters = ComponentRenderHelper.GetCompatibleParameters(this, type);
+        
         parameters.Add(nameof(IMudExFileDisplay.FileDisplayInfos), this);
 
         if (ParametersForSubControls != null)
@@ -527,9 +547,10 @@ public partial class MudExFileDisplay : IMudExFileDisplayInfos
         }
     }
 
-    private async Task ShowInfo()
+    private Task ShowInfo() => ShowInfo(false);
+
+    private async Task ShowInfo(bool showEmptyValues)
     {
-        var selfGenerated = FileInfo == null;
         Stream effectiveStream = ContentStream;
         string size = null;
 
@@ -549,21 +570,51 @@ public partial class MudExFileDisplay : IMudExFileDisplayInfos
             }
         }
 
-        var infoObject = FileInfo ?? new
+        var dict = new Dictionary<string, object>()
         {
-            File = FileName,
-            ContentType,
-            Url = effectiveStream is { Length: > 0 } ? null : Url,
-            Size = size
+            { "File", FileName }, 
+            { "ContentType", ContentType },
+            { "Url", effectiveStream is { Length: > 0 } ? null : Url }, 
+            { "Size", size }
         };
 
+        if (_currentFileDisplay is { Instance: IMudExFileDisplay fileDisplay })
+        {
+            try
+            {
+                var meta = await fileDisplay.FileMetaInformationAsync(this);
+                if (meta != null)
+                    dict.MergeWith(meta);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+        }
+
+        if (showEmptyValues)
+        {
+            foreach (var o in dict.Where(o => o.Value == null))
+                dict[o.Key] = string.Empty;
+        }
+        else
+        {
+            dict = dict.Where(o => o.Value != null).ToDictionary(o => o.Key, o => o.Value);
+        }
+
+        var infoObject = FileInfo ?? ReflectionHelper.CreateTypeAndDeserialize(dict);
+        
         var options = DialogServiceExt.DefaultOptions();
         options.CloseButton = true;
+        options.Resizeable = true;
+        options.DragMode = MudDialogDragMode.Simple;
+        
         await Get<IDialogService>().ShowObject(infoObject, TryLocalize("Info"), Icons.Material.Filled.Info, options, meta =>
         {
-            if (selfGenerated)
-                meta.Properties().Where(p => p.Value == null).Ignore();
+            meta.AllProperties.WrapInMudItem(i => i.xs = 6);
+            meta.Property("Url").WrapInMudItem(i => i.xs = 12);
         });
+        
     }
 
     /// <inheritdoc />