+++ MudBlazor.Extensions/Components/MudExFileDisplayHtml.razor.cs
@@ -65,5 +65,5 @@ public partial class MudExFileDisplayHtml: IMudExFileDisplay
         }
     }
 
-
+    public Task<IDictionary<string, object>> FileMetaInformationAsync(IMudExFileDisplayInfos fileDisplayInfos) => Task.FromResult<IDictionary<string, object>>(null);
 }